services/kitchen-project/requirements.txt
faust-streaming
pydantic